example/requirements.txt
Django=>2.0
django-cors-headers==2.2.0
djangorestframework>=3.6
pdf2image>=1.0.0
Pillow>=5.3.0
